Fix compilation errors about template with gcc and mingw. Fix a lot of warnings about standard c++11 compliance and about mismatching types comparisons

+++ src/examples/Applets/App_MultipleCameras.cpp
@@ -109,7 +109,7 @@ public:
       moneky_w[i]->setEffect(wirefx.get());
     }
 
-    for( int i=0; i<mRenderingTree->subRenderings()->size(); ++i )
+    for( size_t i=0; i<mRenderingTree->subRenderings()->size(); ++i )
     {
       mRenderingTree->subRenderings()->at(i)->as<Rendering>()->renderer()->setFramebuffer( framebuffer.get() );
       mRenderingTree->subRenderings()->at(i)->as<Rendering>()->setRenderer( renderer.get() );
@@ -196,7 +196,7 @@ public:
 
   void mouseDownEvent(EMouseButton, int x, int y)
   {
-    for( int i=0; i<mRenderingTree->subRenderings()->size(); ++i )
+    for( size_t i=0; i<mRenderingTree->subRenderings()->size(); ++i )
     {
       int height = mRenderingTree->subRenderings()->at(i)->as<Rendering>()->renderer()->framebuffer()->height();
       if ( mRenderingTree->subRenderings()->at(i)->as<Rendering>()->camera()->viewport()->isPointInside(x,y,height) )
